requirements.txt
click==6.6
github3.py==1.0.0a4
